+++ Sources/Resources/Shipping/TaskArtifacts/TaskArtifactsClient.swift
@@ -9,7 +9,7 @@ public final class TaskArtifactsClient: Sendable {
         self.httpClient = HTTPClient(config: config)
     }
 
-    /// Adds a new task artifact to a task. | authz_personas=[courier_driver, courier_org_operators, forwarder_org_operators] | (AddTaskArtifactReq) -> (PydanticObjectId)
+    /// Adds a new task artifact to a task. | authz_personas=[driver_for_executor, executor_org_operators, coordinator_org_operators] | (AddTaskArtifactReq) -> (PydanticObjectId)
     ///
     /// - Parameter requestOptions: Additional options for configuring the request, such as custom headers or timeout settings.
     public func addToTaskV1(taskId: String, request: Requests.AddTaskArtifactReq, requestOptions: RequestOptions? = nil) async throws -> String {
@@ -22,7 +22,7 @@ public final class TaskArtifactsClient: Sendable {
         )
     }
 
-    /// Marks a task artifact as COMPLETED. | authz_personas=[courier_driver, courier_org_operators, forwarder_org_operators] | () -> (bool)
+    /// Marks a task artifact as COMPLETED. | authz_personas=[driver_for_executor, executor_org_operators, coordinator_org_operators] | () -> (bool)
     ///
     /// - Parameter requestOptions: Additional options for configuring the request, such as custom headers or timeout settings.
     public func completeV1(taskArtifactId: String, requestOptions: RequestOptions? = nil) async throws -> Bool {
@@ -34,7 +34,7 @@ public final class TaskArtifactsClient: Sendable {
         )
     }
 
-    /// Marks a task artifact as SKIPPED. | authz_personas=[courier_driver, courier_org_operators, forwarder_org_operators] | () -> (bool)
+    /// Marks a task artifact as SKIPPED. | authz_personas=[driver_for_executor, executor_org_operators, coordinator_org_operators] | () -> (bool)
     ///
     /// - Parameter requestOptions: Additional options for configuring the request, such as custom headers or timeout settings.
     public func skipV1(taskArtifactId: String, requestOptions: RequestOptions? = nil) async throws -> Bool {