+++ Sources/Schemas/NotificationEventsForUsersEnum.swift
@@ -13,4 +13,7 @@ public enum NotificationEventsForUsersEnum: String, Codable, Hashable, CaseItera
     case shippingTaskGroupMessage = "shipping.task_group.message"
     case shippingTaskGroupTasksOrderingUpdated = "shipping.task_group.tasks_ordering_updated"
     case shippingTaskGroupFlightInfoUpdated = "shipping.task_group.flight_info_updated"
+    case listingListingDriverInvitedToBid = "listing.listing.driver_invited_to_bid"
+    case operationsCaseOperatorAssigned = "operations.case.operator_assigned"
+    case operationsCaseGroupAssigned = "operations.case.group_assigned"
 }