+++ Tests/Wire/Resources/Shipping/Tasks/TasksClientWireTests.swift
@@ -101,6 +101,28 @@ import Chrt
         try #require(response == expectedResponse)
     }
 
+    @Test func attemptV11() async throws -> Void {
+        let stub = HTTPStub()
+        stub.setResponse(
+            body: Data(
+                """
+                true
+                """.utf8
+            )
+        )
+        let client = ChrtClient(
+            baseURL: "https://api.fern.com",
+            token: "<token>",
+            urlSession: stub.urlSession
+        )
+        let expectedResponse = true
+        let response = try await client.shipping.tasks.attemptV1(
+            taskId: "task_id",
+            requestOptions: RequestOptions(additionalHeaders: stub.headers)
+        )
+        try #require(response == expectedResponse)
+    }
+
     @Test func deleteV11() async throws -> Void {
         let stub = HTTPStub()
         stub.setResponse(